Rename command to `recordLogFile` & outsource method to logger
