+++ package.json
@@ -47,8 +47,8 @@
         "category": "Manim Notebook"
       },
       {
-        "command": "manim-notebook.openLogFile",
-        "title": "Open Log File",
+        "command": "manim-notebook.recordLogFile",
+        "title": "Record Log File",
         "category": "Manim Notebook"
       }
     ],
