+++ src/extension.ts
@@ -5,7 +5,7 @@ import { ManimCell } from './manimCell';
 import { ManimCellRanges } from './manimCellRanges';
 import { previewCode } from './previewCode';
 import { startScene, exitScene } from './startStopScene';
-import { Logger, Window, loggerName } from './logger';
+import { Logger, Window, loggerName, recordLogFile } from './logger';
 
 export function activate(context: vscode.ExtensionContext) {
 	// Trigger the Manim shell to start listening to the terminal
@@ -45,10 +45,10 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 	);
 
-	const openLogFileCommand = vscode.commands.registerCommand(
-		'manim-notebook.openLogFile', async () => {
-			Logger.info("💠 Command requested: Open Log File");
-			openLogFile(context);
+	const recordLogFileCommand = vscode.commands.registerCommand(
+		'manim-notebook.recordLogFile', async () => {
+			Logger.info("💠 Command requested: Record Log File");
+			recordLogFile(context);
 		});
 
 	context.subscriptions.push(
@@ -57,7 +57,7 @@ export function activate(context: vscode.ExtensionContext) {
 		startSceneCommand,
 		exitSceneCommand,
 		clearSceneCommand,
-		openLogFileCommand
+		recordLogFileCommand
 	);
 	registerManimCellProviders(context);
 
@@ -184,50 +184,5 @@ function registerManimCellProviders(context: vscode.ExtensionContext) {
 
 	if (window.activeTextEditor) {
 		manimCell.applyCellDecorations(window.activeTextEditor);
-	}
-}
-
-/**
- * Opens the Manim Notebook log file in a new editor.
- * 
- * @param context The extension context.
- */
-function openLogFile(context: vscode.ExtensionContext) {
-	const logFilePath = vscode.Uri.joinPath(context.logUri, `${loggerName}.log`);
-	window.withProgress({
-		location: vscode.ProgressLocation.Notification,
-		title: "Opening Manim Notebook log file...",
-		cancellable: false
-	}, async (progressIndicator, token) => {
-		await new Promise<void>(async (resolve) => {
-			try {
-				const doc = await vscode.workspace.openTextDocument(logFilePath);
-				await window.showTextDocument(doc);
-			} catch {
-				Window.showErrorMessage("Could not open Manim Notebook log file");
-			} finally {
-				resolve();
-			}
-
-			try {
-				await revealFileInOS(logFilePath);
-			} catch (error: any) {
-				window.showErrorMessage(`Could not open Manim Notebook log file in the`
-					+ ` OS file explorer: ${error?.message}`);
-			}
-		});
-	});
-}
-
-/**
- * Opens a file in the OS file explorer.
- * 
- * @param uri The URI of the file to reveal.
- */
-async function revealFileInOS(uri: vscode.Uri) {
-	if (vscode.env.remoteName === 'wsl') {
-		await vscode.commands.executeCommand('remote-wsl.revealInExplorer', uri);
-	} else {
-		await vscode.commands.executeCommand('revealFileInOS', uri);
 	}
 }
